requirements.txt
requests==2.7.0
dj-database-url==0.3.0
django-crispy-forms==1.5.2
dj-static==0.0.6
Django==1.8.5
django-toolbelt==0.0.1
gunicorn==19.3.
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
django-jsonview==0.5.0
django-dajaxice==0.7
djangorestframework==3.3.0
optional-django==0.1.0
nodejs==0.1.1
numpy==1.9.2
scipy==0.15.1
